fix: -Wsign-compare in multihost fabric and pipeline tests
Fix signed/unsigned comparison warnings in multihost test files:

- test_mesh_device_ttswitch.cpp: auto mesh_id_val=0 → uint32_t
  (compared to *switch_id which is StrongType<uint32_t>)
- mesh_socket_test_context.cpp: world_size cast to uint32_t;
  loop over distributed_context_->size() cast to uint32_t
- intermesh_routing_test_utils.cpp: cast rank() to uint32_t
  before comparing to uint32_t sender_rank
- multihost_fabric_fixtures.hpp: cast global_distributed_context().size()
  to size_t before comparing to vector::size() (4 sites)
- intermesh_routing.cpp: cast rank() to uint32_t for comparisons (2 sites)
- socket_send_recv.cpp: change recv_rank vars from uint32_t to int
  to match Rank = StrongType<int> comparisons
- multi_host_pipeline.cpp: fix int loop var vs size_t (line 83);
  cast my_rank to uint32_t for num_ranks-1 comparison (line 1068)

+++ tests/tt_metal/multihost/fabric_tests/intermesh_routing.cpp
@@ -89,9 +89,9 @@ TEST_F(InterMeshSplit1x2FabricFixture, MultiHopUnicast) {
     constexpr uint32_t num_iterations = 20;
     auto run_send_recv = [&](uint32_t sender_rank, uint32_t recv_rank) {
         for (uint32_t i = 0; i < num_iterations; i++) {
-            if (*(distributed_context->rank()) == sender_rank) {
+            if (static_cast<uint32_t>(*(distributed_context->rank())) == sender_rank) {
                 multihost_utils::run_unicast_sender_step(this, tt::tt_metal::distributed::multihost::Rank{recv_rank});
-            } else if (*(distributed_context->rank()) == recv_rank) {
+            } else if (static_cast<uint32_t>(*(distributed_context->rank())) == recv_rank) {
                 multihost_utils::run_unicast_recv_step(this, tt::tt_metal::distributed::multihost::Rank{sender_rank});
             }
         }
